fixing github actions lint failure
